% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediateP.R
\name{print.mediate}
\alias{print.mediate}
\title{Print mediation results}
\usage{
\method{print}{mediate}(x, ...)
}
\arguments{
\item{x}{An object of class `mediate`.}

\item{...}{other parameters in `print`.}
}
\description{
Print mediation results
}
